+++ tests/test_FindTest.py
@@ -759,7 +759,28 @@ class FindTest(TestCase):
                     year=2007, volume='66', reporter='Cal.Rptr.3d', page='1',
                     metadata={'plaintiff': 'Yield Dynamics, Inc.', 'defendant': 'TEA Systems Corp.', "year": "2007"}
                 ),
-            ])
+            ]),
+            # ignore bad reference matching in plain text
+            ("In Foo v. Bar 1 U.S. 1, ... yadda yadda yadda the case of Foo v. Bar, supra, was affirmed", [
+                case_citation(metadata={'plaintiff': 'Foo', "defendant": "Bar",
+                                        "court": "scotus"}),
+                supra_citation("supra.", metadata={'antecedent_guess': "Bar"})
+            ]),
+            # Dont make second foo a reference citation
+            ('<p>In <i>Foo</i> v. <i>Bar,</i> 1 U.S. 1, ... yadda yadda yadda the case of <i>Foo</i> v. <i>Bar, supra</i><i>,</i> was affirmed.</p>',
+             [
+                case_citation(metadata={'plaintiff': 'Foo', "defendant": "Bar", "court": "scotus"}),
+                supra_citation("supra.", metadata={'antecedent_guess': "Bar"})
+             ], {'clean_steps': ['html', 'inline_whitespace']}),
+            ("""<em>Grant </em>v. <em>Esquire, Inc. </em>(S.D.N.Y. 1973) 367 F.Supp. 876, 879.) ...
+             <em>Grant </em>v. <em>Esquire, Inc., supra, </em>316 F.Supp. at 884.)</p>
+             <p id="b883-5">III. </p> """,
+             [
+                 case_citation(page='876', reporter='F.Supp.', volume='367',
+                               metadata={'plaintiff': 'Grant', "defendant": "Esquire, Inc.", "pin_cite": "879"}),
+                 case_citation(volume="316", reporter="F.Supp.", page="884", short=True,
+                               metadata={'antecedent_guess': "Esquire, Inc., supra", "pin_cite": "884"}),
+             ], {'clean_steps': ['html', 'inline_whitespace']}),
         )
 
         # fmt: on